consumer/requirements.txt
sqlalchemy
psycopg2
pyyaml
kafka-python
